pyproject.toml
[tool.poetry]
name = "car-parking-rest"
version = "0.1.0"
description = "Car parking system to manage space, car and paking status"
authors = ["bibek977 <bhattaraibibek91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "2.32.3"
amqp = "5.2.0"
asgiref = "3.8.1"
async-timeout = "4.0.3"
billiard = "4.2.0"
build = "1.2.1"
cachecontrol = "0.14.0"
celery = "5.4.0"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
crashtest = "0.4.1"
cryptography = "42.0.8"
distlib = "0.3.8"
django = "5.0.7"
django-filter = "24.2"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
dulwich = "0.21.7"
fastjsonschema = "2.20.0"
filelock = "3.15.4"
idna = "3.7"
importlib-metadata = "8.0.0"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jeepney = "0.8.0"
keyring = "24.3.1"
kombu = "5.3.7"
markdown = "3.6"
more-itertools = "10.3.0"
msgpack = "1.0.8"
packaging = "24.1"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
prompt-toolkit = "3.0.47"
ptyprocess = "0.7.0"
pycparser = "2.22"
pyjwt = "2.8.0"
pyproject-hooks = "1.1.0"
python-dateutil = "2.9.0.post0"
rapidfuzz = "3.9.4"
redis = "5.0.7"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.5.1"
tomli = "2.0.1"
tomlkit = "0.13.0"
trove-classifiers = "2024.7.2"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.2"
vine = "5.1.0"
virtualenv = "20.26.3"
wcwidth = "0.2.13"
zipp = "3.19.2"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
faker = "^26.0.0"
toml = "^0.10.2"
sphinx = "^7.4.7"


[tool.poetry.group.dev.dependencies]
redis = "^5.0.7"
celery = "^5.4.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "car_parking_rest.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
